Separate FamplexGraph and file loading utilities
